refactor: Improve home screen pagination and add README
This commit introduces several improvements to the home screen's pagination logic and adds a comprehensive README file for the project.

*   **`README.md`**: Adds a new README file detailing the project's features, architecture, tech stack, and build instructions.
*   **`HomeViewModel.kt`**: Refactors the pagination logic to prevent redundant API calls by canceling previous requests. It also simplifies the logic for checking if the end of the employee list has been reached.
*   **`HomeScreen.kt`**: Improves the pagination trigger by fetching the next page only when the user scrolls to the very last item in the list, providing a more deliberate user experience.

+++ app/src/main/java/com/abhinand/pixbittest/add_employee/data/remote/dto/AddEmployeeRequest.kt
@@ -15,6 +15,6 @@ data class AddEmployeeRequest(
     val profile_pic: File?,
     val resume: File?,
     val contract_period: Int,
-    val total_salary: Double,
+    val total_salary: Int,
     val monthly_payments: List<PaymentDetail>
 )
